run staging in a separate pass
