+++ mlir/test/Dialect/SparseTensor/sparse_vector_mv.mlir
@@ -1,5 +1,3 @@
-// UNSUPPORTED: target={{.*}}
-//
 // RUN: mlir-opt %s -sparse-compiler="vl=8" |  FileCheck %s
 
 #Dense = #sparse_tensor.encoding<{
